+++ modules/frontroute/src/main/scala/io/frontroute/Directives.scala
@@ -13,22 +13,22 @@ trait Directives extends DirectiveApplyConverters {
   def reject: Route = (_, _, _) => Util.rejected
 
   private[frontroute] def extractContext: Directive[RouteLocation] =
-    Directive[RouteLocation](inner => (ctx, previous, state) => inner(ctx)(ctx, previous, state))
+    Directive[RouteLocation](inner => (location, previous, state) => inner(location)(location, previous, state))
 
   private[frontroute] def extract[T](f: RouteLocation => T): Directive[T] =
     extractContext.map(f)
 
   def state[T](initial: => T): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
+      (location, previous, state) => {
         val next = state.enter
         state.getPersistentValue[T](next.path.key) match {
           case None =>
             val newStateValue = initial
-            inner(newStateValue)(ctx, previous, next.setPersistentValue(newStateValue))
+            inner(newStateValue)(location, previous, next.setPersistentValue(newStateValue))
 
           case Some(existing) =>
-            inner(existing)(ctx, previous, next)
+            inner(existing)(location, previous, next)
         }
       }
     )
@@ -36,9 +36,33 @@ trait Directives extends DirectiveApplyConverters {
 
   def signal[T](signal: Signal[T]): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
+      (location, previous, state) => {
         signal.flatMap { extracted =>
-          inner(extracted)(ctx, previous, state.enterAndSet(extracted))
+          inner(extracted)(location, previous, state.enterAndSet(extracted))
+        }
+      }
+    )
+  }
+
+  def memoize[T](retrieve: () => EventStream[T]): Directive[T] = {
+    Directive[T](inner =>
+      (location, previous, state) => {
+        state.async.get((state.path.key, state.data)) match {
+          case Some(value) =>
+            inner(value.asInstanceOf[T])(location, previous, state.enterAndSet(value.asInstanceOf[T]))
+          case _ =>
+            var retrieved = 0
+            retrieve()
+              .filter { _ =>
+                retrieved = retrieved + 1
+                retrieved == 1
+              }
+              .flatMap { retrieved =>
+                val newState = state.copy(
+                  async = state.async.updated((state.path.key, state.data), retrieved)
+                )
+                inner(retrieved)(location, previous, newState.enterAndSet(retrieved))
+              }
         }
       }
     )
@@ -46,9 +70,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def param(name: String): Directive[String] = {
     Directive[String](inner =>
-      (ctx, previous, state) => {
-        ctx.params.get(name).flatMap(_.headOption) match {
-          case Some(paramValue) => inner(paramValue)(ctx, previous, state.enterAndSet(paramValue))
+      (location, previous, state) => {
+        location.params.get(name).flatMap(_.headOption) match {
+          case Some(paramValue) => inner(paramValue)(location, previous, state.enterAndSet(paramValue))
           case None             => Util.rejected
         }
       }
@@ -70,9 +94,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def maybeParam(name: String): Directive[Option[String]] =
     Directive[Option[String]](inner =>
-      (ctx, previous, state) => {
-        val maybeParamValue = ctx.params.get(name).flatMap(_.headOption)
-        inner(maybeParamValue)(ctx, previous, state.enterAndSet(maybeParamValue))
+      (location, previous, state) => {
+        val maybeParamValue = location.params.get(name).flatMap(_.headOption)
+        inner(maybeParamValue)(location, previous, state.enterAndSet(maybeParamValue))
       }
     )
 
@@ -89,13 +113,13 @@ trait Directives extends DirectiveApplyConverters {
   def extractOrigin: Directive[Option[String]] = extract(_.origin)
 
   def provide[L](value: L): Directive[L] =
-    Directive(inner => (ctx, previous, state) => inner(value)(ctx, previous, state.enterAndSet(value)))
+    Directive(inner => (location, previous, state) => inner(value)(location, previous, state.enterAndSet(value)))
 
   def pathPrefix[T](m: PathMatcher[T]): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
-        m(ctx.unmatchedPath) match {
-          case PathMatchResult.Match(t, rest) => inner(t)(ctx.withUnmatchedPath(rest), previous, state.enterAndSet(t))
+      (location, previous, state) => {
+        m(location.unmatchedPath) match {
+          case PathMatchResult.Match(t, rest) => inner(t)(location.withUnmatchedPath(rest), previous, state.enterAndSet(t))
           case _                              => Util.rejected
         }
       }
@@ -104,9 +128,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def testPathPrefix[T](m: PathMatcher[T]): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
-        m(ctx.unmatchedPath) match {
-          case PathMatchResult.Match(t, _) => inner(t)(ctx, previous, state.enterAndSet(t))
+      (location, previous, state) => {
+        m(location.unmatchedPath) match {
+          case PathMatchResult.Match(t, _) => inner(t)(location, previous, state.enterAndSet(t))
           case _                           => Util.rejected
         }
       }
@@ -115,9 +139,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def pathEnd: Directive0 =
     Directive[Unit](inner =>
-      (ctx, previous, state) => {
-        if (ctx.unmatchedPath.isEmpty) {
-          inner(())(ctx, previous, state.enter)
+      (location, previous, state) => {
+        if (location.unmatchedPath.isEmpty) {
+          inner(())(location, previous, state.enter)
         } else {
           Util.rejected
         }
@@ -126,9 +150,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def path[T](m: PathMatcher[T]): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
-        m(ctx.unmatchedPath) match {
-          case PathMatchResult.Match(t, Nil) => inner(t)(ctx.withUnmatchedPath(List.empty), previous, state.enterAndSet(t))
+      (location, previous, state) => {
+        m(location.unmatchedPath) match {
+          case PathMatchResult.Match(t, Nil) => inner(t)(location.withUnmatchedPath(List.empty), previous, state.enterAndSet(t))
           case _                             => Util.rejected
         }
       }
@@ -137,9 +161,9 @@ trait Directives extends DirectiveApplyConverters {
 
   def testPath[T](m: PathMatcher[T]): Directive[T] = {
     Directive[T](inner =>
-      (ctx, previous, state) => {
-        m(ctx.unmatchedPath) match {
-          case PathMatchResult.Match(t, Nil) => inner(t)(ctx, previous, state.enterAndSet(t))
+      (location, previous, state) => {
+        m(location.unmatchedPath) match {
+          case PathMatchResult.Match(t, Nil) => inner(t)(location, previous, state.enterAndSet(t))
           case _                             => Util.rejected
         }
       }
@@ -154,17 +178,17 @@ trait Directives extends DirectiveApplyConverters {
     )
 
   def debug(message: Any, optionalParams: Any*)(subRoute: Route): Route =
-    (ctx, previous, state) => {
+    (location, previous, state) => {
       dom.console.debug(message, optionalParams: _*)
-      subRoute(ctx, previous, state)
+      subRoute(location, previous, state)
     }
 
-  def concat(routes: Route*): Route = (ctx, previous, state) => {
+  def concat(routes: Route*): Route = (location, previous, state) => {
     def findFirst(rs: List[(Route, Int)]): EventStream[RouteResult] =
       rs match {
         case Nil => Util.rejected
         case (route, index) :: tail =>
-          route(ctx, previous, state.enterConcat(index)).flatMap {
+          route(location, previous, state.enterConcat(index)).flatMap {
             case complete: RouteResult.Complete => EventStream.fromValue(complete)
             case RouteResult.Rejected           => findFirst(tail)
           }