+++ project/Dependencies.scala
@@ -31,6 +31,7 @@ object Dependencies {
       "io.laminext" %%% "ui"                     % DependencyVersions.laminext,
       "io.laminext" %%% "tailwind"               % DependencyVersions.laminext,
       "io.laminext" %%% "util"                   % DependencyVersions.laminext,
+      "io.laminext" %%% "fetch"                  % DependencyVersions.laminext,
       "io.laminext" %%% "tailwind-default-theme" % DependencyVersions.laminext
     )
   }